+++ CHANGELOG.md
@@ -10,6 +10,28 @@ and this project adheres to [Semantic Versioning][].
 
 ## [Unreleased]
 
-### Added
+## [v0.2.0]
 
-- Basic tool, preprocessing and plotting functions
+### Added
+- PyPI badge and installation instructions [#6]
+- Added two examples for data integration: one for scRNA-seq data [#13], [#14], [#15], [#16] and one for spatial data (seqFISH) [#17]
+- Added a count parameter [#9]
+
+## [v0.1.3]
+
+### Changed
+- Attempt again to publish to PyPI
+
+## [v0.1.2]
+
+### Changed
+- Attempt to publish to PyPI
+
+## [v0.1.1]
+
+### Added
+- Set up pre-commit.ci and codecov [#4]
+- Release on PyPI
+
+## [v0.1.0]
+Basic version of this package to create a `submit.sh` script and initialize a hyperparameter sweep on W&B.